✨ translate api 정리 #15
