+++ eight/src/pages/docent/DocentExp.js
@@ -9,7 +9,6 @@ import typo from "../../styles/typo";
 import PartialInfo from "../../components/docent/PartialInfo";
 import { useTranslation } from "react-i18next";
 import { useParams } from "react-router-dom";
-import axios from "axios";
 
 const DocentExp = ({ artInfo }) => {
   const { t } = useTranslation();
@@ -28,34 +27,6 @@ const DocentExp = ({ artInfo }) => {
   const handleImageLoad = (event) => {
     const imgElement = event.target;
     setArtImageHeight(imgElement.height);
-  };
-
-  //번역
-  const translate = async (q, lng) => {
-    const tData = JSON.stringify({
-      q: q,
-      target: lng,
-    });
-    let config = {
-      method: "post",
-      maxBodyLength: 500,
-      url: `https://translation.googleapis.com/language/translate/v2?key=${process.env.REACT_APP_GOOGLE_API_KEY}`,
-      headers: {
-        "Content-Type": "application/json",
-      },
-      data: tData,
-    };
-    return await axios
-      .request(config)
-      .then((response) => {
-        return JSON.stringify(response.data.data.translations[0].translatedText)
-          .replace(/"/g, "")
-          .replace(/&lt;/g, "<")
-          .replace(/&gt;/g, ">");
-      })
-      .catch((error) => {
-        console.log(error);
-      });
   };
 
   return (
@@ -95,14 +66,7 @@ const DocentExp = ({ artInfo }) => {
           // 부분별 해설 탭
           <>
             {artInfo[params].quest.map((data, idx) => {
-              return (
-                <PartialInfo
-                  key={idx}
-                  artInfo={data}
-                  translate={translate}
-                  t={t}
-                />
-              );
+              return <PartialInfo key={idx} artInfo={data} t={t} />;
             })}
           </>
         ) : (
